Break Product into smallers components
