+++ src/modules/products-page/product-id-page/ProductDetail/responsive.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/variables";
+@import "../../../../scss/variables";
 
 @media screen and (max-width: 1200px) {
   .product-quantity-selector {
