+++ src/modules/products-page/product-id-page/ProductDetail/style.scss
@@ -1,0 +1,124 @@
+@import "../../../../scss/variables";
+
+.single-product-detail {
+    margin-top: 2rem;
+    display: grid;
+    grid-template-columns: repeat(7, 1fr);
+}
+
+.product-image-list {
+    width: 100%;
+    > ul {
+        list-style: none;
+        padding: 0;
+    }
+}
+.product-image-list-item {
+    max-width: 100%;
+    max-height: 100%;
+    padding-bottom: 1rem;
+    > * {
+        max-width: 100%;
+        max-height: 100%;
+    }
+    > *:hover {
+        border: 1px solid red;
+        cursor: pointer;
+    }
+}
+
+.product-image-highlight {
+    position: relative;
+    grid-column-start: 2;
+    grid-column: span 3;
+    display: flex;
+    justify-content: center;
+    max-width: 100%;
+    max-height: 100%;
+    margin-left: 1rem;
+    > img {
+        max-width: 100%;
+        max-height: 100%;
+    }
+}
+.image-highlight {
+    position: absolute;
+}
+.image-selected {
+    border: 1px solid red;
+}
+.highlighted {
+    opacity: 1;
+    transition: 0.5s;
+}
+.transparent {
+    opacity: 0;
+    transition: 0.5s;
+}
+.product-description {
+    grid-column: span 3;
+    padding-left: 3rem;
+    > h1 {
+        margin: 0;
+        margin-bottom: 1rem;
+    }
+}
+.product-benefit-wrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: $light-sliver;
+    margin-top: 3rem;
+    > div {
+        padding: 1rem;
+        > h5 {
+            margin: 0;
+        }
+    }
+}
+
+.product-price {
+    margin: 2rem 0;
+    font-size: 40px;
+    color: red;
+}
+
+.product-quantity-selector {
+    > p {
+        margin: 0;
+    }
+    display: flex;
+    justify-content: space-between;
+}
+.quantity-selector-wrapper {
+    display: flex;
+    justify-content: space-between;
+    font-size: 20px;
+    width: 50%;
+}
+.quantity-selector {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border: 0.5px solid $sliver;
+
+    .plus-minus-button {
+        padding: 0 1rem;
+    }
+    .quantity-value {
+        font-size: 20px;
+    }
+}
+.product-addtocart-wrapper {
+    display: flex;
+    width: 50%;
+    justify-content: space-around;
+    height: auto;
+    align-items: center;
+}
+
+.product-favorite-icon {
+    color: $main-color;
+    border: 0.5px solid $sliver;
+    padding: 0.5rem 0.3rem;
+}
